feat: create play and leaderboard buttons && change some texting
