+++ buttons/leaderboard.js
@@ -1,0 +1,5 @@
+const leaderboard = async (interaction) => {
+  console.log('leaderboard', interaction);
+};
+
+module.exports = leaderboard;
